+++ FinalEngine.Editor.Desktop/App.xaml.cs
@@ -12,7 +12,7 @@ using FinalEngine.Editor.Common.Extensions;
 using FinalEngine.Editor.Common.Services.Application;
 using FinalEngine.Editor.Common.Services.Environment;
 using FinalEngine.Editor.Desktop.Factories;
-using FinalEngine.Editor.Desktop.Interactions;
+using FinalEngine.Editor.Desktop.Services.Actions;
 using FinalEngine.Editor.Desktop.Views;
 using FinalEngine.Editor.Desktop.Views.Dialogs.Layout;
 using FinalEngine.Editor.ViewModels;
@@ -24,6 +24,7 @@ using FinalEngine.Editor.ViewModels.Docking.Tools.Projects;
 using FinalEngine.Editor.ViewModels.Docking.Tools.Scenes;
 using FinalEngine.Editor.ViewModels.Factories;
 using FinalEngine.Editor.ViewModels.Interactions;
+using FinalEngine.Editor.ViewModels.Services.Actions;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
