+++ FinalEngine.Tests/Editor/ViewModels/Dialogs/Layouts/SaveWindowLayoutViewModelTests.cs
@@ -8,6 +8,8 @@ using System;
 using FinalEngine.Editor.ViewModels.Dialogs.Layout;
 using FinalEngine.Editor.ViewModels.Factories;
 using FinalEngine.Editor.ViewModels.Interactions;
+using FinalEngine.Editor.ViewModels.Services.Actions;
+using FinalEngine.Editor.ViewModels.Services.Layout;
 using Moq;
 using NUnit.Framework;
