+++ FinalEngine.Tests/Editor/ViewModels/Docking/DockViewModelTests.cs
@@ -13,7 +13,7 @@ using FinalEngine.Editor.ViewModels.Docking.Tools.Inspectors;
 using FinalEngine.Editor.ViewModels.Docking.Tools.Projects;
 using FinalEngine.Editor.ViewModels.Docking.Tools.Scenes;
 using FinalEngine.Editor.ViewModels.Factories;
-using FinalEngine.Editor.ViewModels.Interactions;
+using FinalEngine.Editor.ViewModels.Services.Layout;
 using Moq;
 using NUnit.Framework;
